requirements.txt
tensorflow==2.15.0
opencv-python==4.8.0
numpy==1.24.0
matplotlib==3.7.0
pandas==2.0.0
seaborn==0.12.0
keras-tuner==1.4.0
pytest==7.4.0
PyYAML==6.0
jupyter==1.0.0  # For running notebooks in README instructions
# Install: pip install -r requirements.txt (tested Sep 27, 2025)